refactor(formatters): split 599-line console.ts into per-mode files
console.ts mixed five concerns (shared analysis pipeline, summary,
chronological list, follow stream, JSON) in one 599-line file. Extract
each mode into its own file under formatters/console/ alongside a
shared.ts that holds the types, LEVEL_MAP, and the analyze/dedupe
helpers. The barrel at formatters/console.ts re-exports the public
surface and keeps the formatConsole dispatcher, so consumers don't
change. All 650 tests pass. Refs #223.
